% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{alpha_centrality}
\alias{alpha_centrality}
\title{Get Alpha Centrality}
\usage{
alpha_centrality(A, a)
}
\arguments{
\item{A}{Square adjacency matrix}

\item{a}{penalty term}
}
\description{
Get Alpha Centrality
}
