% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_adj_mat}
\alias{create_adj_mat}
\title{Create adjacency matrix from given  data. Rows are typically voters
and columns are posters. Nrow and and ncol should be identical.
Matrix is one if an edge voter-poster exists.}
\usage{
create_adj_mat(x, y, nrow, ncol = nrow)
}
\arguments{
\item{nrow}{number of rows of resulting table}

\item{ncol}{number of columns of resulting table}

\item{voter}{vector of voters}

\item{poster}{poster vector of posters}
}
\value{
sparse adjency matrix
}
\description{
Vector of voters and posters should be adjusted to represent numbers from
1 to #unique users
}
