% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_adj_mat_time}
\alias{create_adj_mat_time}
\title{Create a list of adjacency matrix of a time variable. Creates a list of
matrices (A1, A2, ..., Ak)}
\usage{
create_adj_mat_time(df_input, time_var, x = "ID_VoteUser", y = "ID_PostUser")
}
\arguments{
\item{time_var}{column name of time variable to use}

\item{x}{outgoing column name}

\item{y}{ingoing column name}

\item{df}{inputs data frame}
}
\value{
list of:
 "matrix" - list of adjacency matrix at each time point
 "mapping" - mapping for the user ids using mapping[user_id]
}
\description{
Create a list of adjacency matrix of a time variable. Creates a list of
matrices (A1, A2, ..., Ak)
}
