% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_relevant_data.R
\name{create_user_vote_user_data}
\alias{create_user_vote_user_data}
\title{create user-vote-user-data}
\usage{
create_user_vote_user_data(
  df_votes,
  df_posts,
  channels = NULL,
  save_path = NULL
)
}
\arguments{
\item{df_votes}{data of votes with at least "ID_Posting" and "ID_CommunityIdentity" columns}

\item{df_posts}{data of posts with at least "ID_Posting" and "ID_CommunityIdentity" columns}

\item{channels}{default NULL (all channels), specify specific article channels}

\item{save_path}{default NULL, if not NULL: save data under this path}
}
\value{
data of class data.table
}
\description{
create user-vote-user-data
}
