% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{dcm_simple}
\alias{dcm_simple}
\title{Dynamic Communicability Matrix (dcm)}
\usage{
dcm_simple(B)
}
\arguments{
\item{B}{List of alpha centrality matrices}
}
\value{
List of dynamic communicability matrices
}
\description{
Input list of alpha centralities created using the function \code{\link{store_alpha}}
and calculates dynamic communicability matrices as described in "A matrix
iteration for dynamic network summaries" by Grindrod, P. and Higham.
}
