% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{get_alphas}
\alias{get_alphas}
\title{Stores list of alpha centralities of list of adjacency matrices}
\usage{
get_alphas(A, a, path = NULL)
}
\arguments{
\item{A}{is a list of adjacency matrices. Each adjacency matrix must be
square}

\item{a}{penalty term}
}
\value{
List of Alpha Centralities
}
\description{
Stores list of alpha centralities of list of adjacency matrices
}
