% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{get_upperbound_a}
\alias{get_upperbound_a}
\title{Find upper bound for alpha}
\usage{
get_upperbound_a(A, exact = FALSE)
}
\arguments{
\item{A}{list of adjacency matrices}

\item{exact}{boolean, if TRUE the exact eigenvalues are calculated, else only the upper
bound of the largest eigenvalue}
}
\value{
upper bound of parameter a
}
\description{
Input list of adjacency matrices to calculate the upper bound of parameter a
through the spectral radius of the adjacency matrices.
WARNING: can be REALLY slow!!!
May not be recommended at all, if matrix is too large or eigenvalue can not be found.
An upper bound of largest eigenvalue or spectral radius should be used instead
hard to find one for our case: non-negative and asymmetric matrix (+non-regular graph)
used Theorem 2.1 from https://www.sciencedirect.com/science/article/pii/S0024379513005405
calculating an upper bound of spectral radius.
}
