% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{get_users}
\alias{get_users}
\title{Find ntop users}
\usage{
get_users(rdcm, type = "broadcast", ntop = 20, mappings)
}
\arguments{
\item{rdcm}{List of running dynamic communicability matrices}

\item{type}{either "broadcast" or "receive"}

\item{ntop}{number of n top users being returned}

\item{mappings}{vector of mapping between "ID" and "Node" (an outcome of the function
\code{\link{create_adj_mat_time}})}
}
\value{
List of ntop users for each time point
}
\description{
Return TOP N users regarding the dynamic broadcast or receive communicabilities
for each time point
Input list of running dynamic communicability matrices (an outcome of the function
\code{\link{rdcm_simple}})
}
