% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{rdcm_simple}
\alias{rdcm_simple}
\title{Running Dynamic Communicability Matrix (RDCM)}
\usage{
rdcm_simple(B, b, dt)
}
\arguments{
\item{B}{List of alpha centrality matrices}

\item{b}{parameter for downweighting older walks}

\item{dt}{vector of time differences between two timepoints}
}
\value{
List of running dynamic communicability matrices
}
\description{
Input list of alpha centralities created using the function \code{\link{store_alpha}}
and parameter b to calculate running dynamic communicability matrices as described
in "A matrix iteration for dynamic network summaries" by Grindrod, P. and Higham.
}
