% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_csv_data}
\alias{read_csv_data}
\title{To read all relevant files csv files}
\usage{
read_csv_data(
  user = "tobias",
  data_names = c("Following", "Votes", "Postings"),
  save_as_fst = FALSE,
  return_df = TRUE
)
}
\arguments{
\item{user}{name of user}

\item{data_names}{data to be loaded}

\item{save_as_fst}{default FALSE, whether to save data in fst format or not
fst-format very flexible: read specific rows and columns
see more: http://www.fstpackage.org/}

\item{return_df}{returns data only if true.}
}
\value{
data of class data.table in a list
}
\description{
To read all relevant files csv files
}
